chore: move router v2 to new community schema, use lerna and update dependencies
