+++ package.json
@@ -2,17 +2,10 @@
   "name": "root",
   "private": true,
   "devDependencies": {
-    "@stencil/core": "^2.1.0-2",
-    "@types/jest": "^26.0.12",
-    "jest": "^26.5.3",
-    "jest-cli": "^26.5.3",
-    "np": "^6.5.0",
-    "prettier": "^2.1.2",
-    "rollup": "^2.32.0",
-    "typescript": "^4.0.3"
+    "lerna": "^5.2.0",
+    "@stencil/core": "^2.17.1"
   },
-  "repository": {
-    "type": "git",
-    "url": "git://github.com/ionic-team/stencil-router-v2.git"
-  }
+  "workspaces": [
+    "packages/*"
+  ]
 }
