+++ packages/demo/src/router.ts
@@ -1,0 +1,2 @@
+import { createStaticRouter } from '@stencil-community/router';
+export const Router = createStaticRouter();
