+++ packages/demo/stencil.config.ts
