+++ packages/demo/tsconfig.json
