+++ packages/router/src/static-router.ts
