+++ packages/router/src/utils/path-to-regex.ts
