+++ packages/router/src/utils/server-static-data-proxy.ts
