+++ packages/router/tsconfig.json
@@ -16,6 +16,8 @@
     "jsx": "react",
     "jsxFactory": "h"
   },
-  "files": ["src/index.ts"],
+  "include": [
+    "src"
+  ],
   "exclude": ["node_modules", "src/**/*.test.ts"]
 }
