--- test-app/src/router.ts
@@ -1,2 +1,0 @@
-import { createStaticRouter } from 'stencil-router-v2';
-export const Router = createStaticRouter();
